+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Make output simpler and more consistent.
 - Limit signal handling to HUP, INT, TERM so that the default
   behavior is used for any other signals (STOP, CONT, etc).
+- Reap all child processes.
 
 # v0.2 (2017-11-15)